% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerterm.R
\name{create_docker_terminal}
\alias{create_docker_terminal}
\title{Create a new terminal window with a R running in a Docker container}
\usage{
create_docker_terminal(id, image = TRUE, entrypoint = "R")
}
\arguments{
\item{id}{character. Docker image id to use for the new container.}
}
\value{
The terminal id for the created terminal.
}
\description{
\code{create_docker_terminal} creates a docker container running R in a new
terminal window. Code from the RStudio source pane can be sent to this
container with \strong{CTRL/CMD + ALT + ENTER}
}
