% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerterm.R
\docType{package}
\name{dockerterm}
\alias{dockerterm}
\alias{dockerterm-package}
\alias{dockerterm}
\title{dockerterm: Run R code in arbitrary docker containers from within RStudio}
\usage{
dockerterm()
}
\description{
This package provides an RStudio Addin that makes it easy to create a
docker container in the terminal pane and then run R code from a source
file in RStudio within that docker container. There are several inherint
limitations to this model. Namely, RStudio isn't connected to the R session
in Docker. Therefore, code completion in RStudio doesn't map to the R
environment within Docker and plots rendered within the Docker container
don't show up in RStudio. For a full fledged "R in Docker Experience" it's
better to run RStudio Server from a Docker Container. However, this Addin
provides the ability to quickly run R code against an arbitrary R backend
within a specified Docker container.

\code{dockerterm} runs a shiny gadget that allows the user to specify a Docker
container or image to run in the terminal window. The Docker container must
have R already installed as this is the command invoked when the image is run.
If the specified image is not found on the users machine, it will be downloaded
from Dockerhub and run.
}
